src/requirements.txt
aiofiles==22.1.0
aiohttp==3.8.3
aiosignal==1.2.0
aiosqlite==0.17.0
anyio==3.6.2
async-timeout==4.0.2
asyncpg==0.26.0
attrs==22.1.0
bcrypt==4.0.1
black==22.10.0
certifi==2022.9.24
charset-normalizer==2.1.1
click==8.1.3
exceptiongroup==1.0.0rc9
frozenlist==1.3.1
h11==0.12.0
httpcore==0.15.0
httptools==0.5.0
httpx==0.23.0
idna==3.4
iniconfig==1.1.1
iso8601==1.1.0
multidict==6.0.2
mypy-extensions==0.4.3
packaging==21.3
passlib==1.7.4
pathspec==0.10.1
platformdirs==2.5.2
pluggy==1.0.0
pycryptodome==3.15.0
pydantic==1.10.2
PyJWT==2.6.0
pyparsing==3.0.9
pypika-tortoise==0.1.6
pytest==7.2.0
pytz==2022.5
PyYAML==6.0
rfc3986==1.5.0
sanic==22.9.0
sanic-ext==22.9.0
sanic-jwt==1.8.0
sanic-routing==22.8.0
sanic-testing==22.9.0
sniffio==1.3.0
tomli==2.0.1
tortoise-orm==0.19.2
typing_extensions==4.4.0
ujson==5.5.0
uvloop==0.17.0
websockets==10.3
yarl==1.8.1
